backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
channels==4.0.0
charset-normalizer==3.1.0
Django==4.2
django-nextjs==2.2.3
django-ninja==0.22.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==1.10.8
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.6.3
urllib3==2.0.2
websockets==11.0.3
yarl==1.9.2
